+++ TinkoffASDKCore/TinkoffASDKCore/API/RequestData/InitAddCardData.swift
@@ -1,0 +1,42 @@
+//
+//
+//  InitAddCardData.swift
+//
+//  Copyright (c) 2021 Tinkoff Bank
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+
+import Foundation
+
+public struct InitAddCardData: Codable {
+    public var checkType: String
+    public var customerKey: String
+
+    public enum CodingKeys: String, CodingKey {
+        case checkType = "CheckType"
+        case customerKey = "CustomerKey"
+    }
+
+    public init(from decoder: Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+        checkType = try container.decode(String.self, forKey: .checkType)
+        customerKey = try container.decode(String.self, forKey: .customerKey)
+    }
+
+    public init(with checkType: String, customerKey: String) {
+        self.checkType = checkType
+        self.customerKey = customerKey
+    }
+}
